+++ test/unit/lib/inspectHandlerTests.js
@@ -817,6 +817,12 @@ describe('inspectHandler', () => {
             '/tm/net/routing/bgp/~Common~exampleBGP/neighbor': [
                 {
                     name: '10.1.1.2',
+                    addressFamily: [
+                        {
+                            name: 'ipv4',
+                            asOverride: 'enabled'
+                        }
+                    ],
                     ebgpMultihop: 2,
                     peerGroup: 'Neighbor_IN',
                     unwanted: 1
@@ -1997,6 +2003,12 @@ describe('inspectHandler', () => {
                             neighbors: [
                                 {
                                     address: '10.1.1.2',
+                                    addressFamilies: [
+                                        {
+                                            internetProtocol: 'ipv4',
+                                            asOverrideEnabled: true
+                                        }
+                                    ],
                                     ebgpMultihop: 2,
                                     peerGroup: 'Neighbor_IN'
                                 }